requirements.txt
requests
PyYAML
python-Levenshtein
flask
gunicorn
pytz
asyncio
tapo
python-nmap
sqlalchemy